Refactor: trying to bring some order to the chaos
