+++ obs_tools/__init__.py
@@ -1,12 +1,14 @@
 from config import AudioMode, config
 
 if config.obs_integration:
-    from .parse_map_loading_screen import LoadingScreenScanner as GameStartedScanner
+    from .events.parse_map_loading_screen import (
+        LoadingScreenScanner as GameStartedScanner,
+    )
 else:
-    from .sc2client import ClientAPIScanner as GameStartedScanner
+    from .events.clientapi import ClientAPIScanner as GameStartedScanner
 
 
 if config.audiomode in [AudioMode.full, AudioMode.voice_in]:
-    from .wakeword import WakeWordListener as WakeListener
+    from .events.wakeword import WakeWordListener as WakeListener
 else:
-    from .wakekey import WakeKeyListener as WakeListener
+    from .events.wakekey import WakeKeyListener as WakeListener
