+++ obs_tools/events/wakekey.py
@@ -5,7 +5,7 @@ import keyboard
 
 from config import config
 
-from .types import WakeResult
+from obs_tools.events.types import WakeResult
 
 log = logging.getLogger(f"{config.name}.{__name__}")
