+++ obs_tools/io/rich_log.py
@@ -27,6 +27,9 @@ class LogStatus(Status):
         self.update(self.buffer)
 
 
+# If log messages are generated by these methods, they will trigger a rich
+# status. A status will be rendered with a spinner and the message will be
+# streamed to the status.
 STATUS_METHODS = ["converse", "transcribe", "playrich", "say"]
 
 
@@ -79,7 +82,8 @@ def get_emoji(name: str, funcName: str) -> str:
 class TwitchObsLogHandler(Handler):
     """Custom log handler for rich console output.
 
-    We are (mis)using the logging system to provide a pretty interface which can be embedded in OBS.
+    We are (mis)using the logging system to provide a pretty interface
+    for text mode and which can be embedded in OBS.
     """
 
     _status_methods: Dict[str, LogStatus] = {}
@@ -91,6 +95,21 @@ class TwitchObsLogHandler(Handler):
         self.fqn = None
 
     def emit(self, record: LogRecord) -> None:
+        """Emit a log record to the rich console.
+
+        Determine which emoji to use at the start of each log line.
+        Determine if the message is streamed ("flushed") in which case it is not
+        printed with a line ending.
+
+        Check if the message should cancel any running statuses.
+
+        If the message is a status message, start a new status and record the status
+        state. If a status from the logging method is already active, stream the
+        message to the status.
+        If a status from a different method is active, stop it and start a new one.
+
+        If the message is not a status message, print the message with the emoji and style.
+        """
         if record.levelno < logging.INFO:
             return
 
@@ -162,6 +181,12 @@ class TwitchObsLogHandler(Handler):
             f.write(f"{time.time()}")
 
     def check_stop(self, record: LogRecord) -> bool:
+        """Determine if we need to stop a status.
+
+        To stop a streamed status, the caller needs to log the fully buffered message again.
+        The buffer will then stop the status and the log message will be printed instead.
+        When the user logs a message, always cancel all statuses.
+        """
         for status in self._status_methods.values():
             if status.status == record.msg:
                 return True
