+++ obs_tools/lib/battlenet.py
