+++ obs_tools/lib/sc2pulse.py
@@ -11,8 +11,8 @@ import httpx
 from pydantic import BaseModel
 
 from config import config
-from obs_tools.battlenet import toon_handle_from_id
-from obs_tools.types import Race as GameInfoRace
+from obs_tools.lib.battlenet import toon_handle_from_id
+from obs_tools.lib.sc2client import Race as GameInfoRace
 from replays.util import convert_enum, is_barcode
 
 log = logging.getLogger(f"{config.name}.{__name__}")
